pyproject.toml
[project]
name = "deepbrowser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.63",
    "deepagents>=0.2.0",
    "langchain>=1.0.2",
    "langchain-aws>=1.0.0",
    "langchain-core>=1.0.1",
    "mypy>=1.18.2",
    "nest-asyncio>=1.6.0",
    "pillow>=12.0.0",
    "playwright>=1.55.0",
    "pydantic>=2.12.3",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "ruff>=0.14.2",
]

[build-system]
requires = ["uv-build"]
build-backend = "uv_build"

[project.scripts]
test = "deepbrowser.test:main"

[tool.ruff]
target-version = "py312"

# Directories Ruff should lint
src = ["src"]

line-length = 100

# Enable autofix for common issues
fix = true
show-fixes = true

lint.select = [
    "F",   # pyflakes
    "E",   # pycodestyle errors
    "I",   # isort
    "W"
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
docstring-code-format = true
skip-magic-trailing-comma = false

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F401"]

[tool.ruff.lint.isort]
# Keep imports clean and deterministic
known-first-party = ["deepbrowser"]
combine-as-imports = true
force-single-line = false
lines-after-imports = 2


[tool.mypy]
python_version = "3.12"
mypy_path = "src"
plugins = ["pydantic.mypy"]

# Start strict, then relax per-module below
strict = true

[[tool.mypy.overrides]]
module = "deepagents.*"
ignore_missing_imports = true

